# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Academic Workspace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me/profile:
    get:
      operationId: getMyProfile
      tags: [profile]
      summary: Get current user's public profile
      description: Returns display name, avatar URL, email, and account info.
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
    patch:
      operationId: updateMyProfile
      tags: [profile]
      summary: Update current user's profile
      description: Update display name and/or avatar URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
        "400":
          description: Invalid input
  /users/me/avatar:
    post:
      operationId: uploadMyAvatar
      tags: [profile]
      summary: Upload avatar image
      description: Upload a profile picture to Supabase Storage. Accepts JPEG, PNG, or WebP up to 5MB.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AvatarUploadRequest"
      responses:
        "200":
          description: Avatar uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvatarUploadResponse"
        "401":
          description: Unauthorized
        "400":
          description: Invalid file type or size
  /users/me/account:
    delete:
      operationId: deleteMyAccount
      tags: [profile]
      summary: Delete current user account
      description: Permanently deletes the user account and all associated data. Requires password confirmation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteAccountRequest"
      responses:
        "200":
          description: Account deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteAccountResponse"
        "401":
          description: Unauthorized or invalid password
        "400":
          description: Invalid request
        "500":
          description: Server error

  # ── Document Templates ───────────────────────────────────────────────────
...
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, email, displayName, avatarUrl, isOwner, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        displayName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        isOwner:
          type: boolean
        referralCode:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    UpdateProfileRequest:
      type: object
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 100
        avatarUrl:
          type: string
          format: uri
    AvatarUploadRequest:
      type: object
      required: [base64Content, filename]
      properties:
        base64Content:
          type: string
          description: Base64-encoded image content (max 5MB, JPEG/PNG/WebP)
        filename:
          type: string
          description: Original filename (e.g. "avatar.jpg")
    AvatarUploadResponse:
      type: object
      required: [avatarUrl]
      properties:
        avatarUrl:
          type: string
          format: uri
          description: Public URL of the uploaded avatar
    DeleteAccountRequest:
      type: object
      required: [password]
      properties:
        password:
          type: string
          description: User's current password to confirm deletion
    DeleteAccountResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ── Account References ───────────────────────────────────────────────────────
...